Refactor out TerraformSource struct and functions so we can use it in config package without circular dependency
